fix(gates): split gc/layout.rs slot visitors under the 2000-line cap
